requirements.txt
Flask==3.1.0
Flask-Bcrypt==1.0.1
face-recognition==1.3.0
mysql-connector-python==9.2.0
numpy==1.26.4
opencv-python==4.10.0.84
PyJWT==2.10.1
python-dotenv==1.0.1
